+++ src/image/SkImage.cpp
@@ -224,23 +224,34 @@ sk_sp<SkShader> SkImage::makeRawShader(SkTileMode tmx, SkTileMode tmy,
                                   sampling, localMatrix);
 }
 
-sk_sp<SkData> SkImage::encodeToData(SkEncodedImageFormat type, int quality) const {
-    // Context TODO: Elevate GrDirectContext requirement to public API.
-    auto dContext = as_IB(this)->directContext();
+sk_sp<SkData> SkImage::encodeToData(GrDirectContext* context, SkEncodedImageFormat type,
+                                    int quality) const {
     SkBitmap bm;
-    if (as_IB(this)->getROPixels(dContext, &bm)) {
+    if (as_IB(this)->getROPixels(context, &bm)) {
         return SkEncodeBitmap(bm, type, quality);
     }
     return nullptr;
 }
 
-sk_sp<SkData> SkImage::encodeToData() const {
+sk_sp<SkData> SkImage::encodeToData(GrDirectContext* context) const {
     if (auto encoded = this->refEncodedData()) {
         return encoded;
     }
 
-    return this->encodeToData(SkEncodedImageFormat::kPNG, 100);
+    return this->encodeToData(context, SkEncodedImageFormat::kPNG, 100);
 }
+
+#ifndef SK_IMAGE_READ_PIXELS_DISABLE_LEGACY_API
+sk_sp<SkData> SkImage::encodeToData(SkEncodedImageFormat type, int quality) const {
+    auto dContext = as_IB(this)->directContext();
+    return this->encodeToData(dContext, type, quality);
+}
+
+sk_sp<SkData> SkImage::encodeToData() const {
+    auto dContext = as_IB(this)->directContext();
+    return this->encodeToData(dContext);
+}
+#endif
 
 sk_sp<SkData> SkImage::refEncodedData() const {
     return sk_sp<SkData>(as_IB(this)->onRefEncoded());