+++ modules/canvaskit/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Skottie TransformValue accessors for dynamic layer transforms.
 
+### Changed
+ - `Image.encodeToData` now makes use of the GPU context more consistently.
+
 ## [0.38.0] - 2023-01-12
 
 ### Changed